{"types":"C","runs":34}

+++ problem49.java
@@ -3,9 +3,9 @@ import acm.program.ConsoleProgram;
 public class problem49 extends ConsoleProgram{
     public void run(){
     	String s = readLine("Enter the text: ");
-    	double ans = 0;
+    	int ans = 0;
     	for(int i = s.length() - 1; i >= 0; i--){
-    		ans = ans + Math.pow(10, i) *((int)s.charAt(s.length() - i - 1) - 48);  
+    		ans = ans + (int)(Math.pow(10, i) *((int)s.charAt(s.length() - i - 1)) - 48);  
     	}
     	print(ans);
     }
